# [Perf] Memoize relative paths to speed up builds