--- a/lib/optimize/AggressiveSplittingPlugin.js
+++ b/lib/optimize/AggressiveSplittingPlugin.js
@@ -56,7 +56,7 @@ class AggressiveSplittingPlugin {
 
 						const nameToModuleMap = new Map();
 						chunk.forEachModule(m => {
-							const name = identifierUtils.makePathsRelative(compiler.context, m.identifier());
+							const name = identifierUtils.makePathsRelative(compiler.context, m.identifier(), compilation.cache);
 							nameToModuleMap.set(name, m);
 						});
 
@@ -127,7 +127,7 @@ class AggressiveSplittingPlugin {
 							newChunk.origins = chunk.origins.map(copyWithReason);
 							chunk.origins = chunk.origins.map(copyWithReason);
 							compilation._aggressiveSplittingSplits = (compilation._aggressiveSplittingSplits || []).concat({
-								modules: newChunk.mapModules(m => identifierUtils.makePathsRelative(compiler.context, m.identifier()))
+								modules: newChunk.mapModules(m => identifierUtils.makePathsRelative(compiler.context, m.identifier(), compilation.cache))
 							});
 							return true;
 						} else {
@@ -144,7 +144,7 @@ class AggressiveSplittingPlugin {
 					if(chunk.hasEntryModule()) return;
 					const size = chunk.size(this.options);
 					const incorrectSize = size < minSize;
-					const modules = chunk.mapModules(m => identifierUtils.makePathsRelative(compiler.context, m.identifier()));
+					const modules = chunk.mapModules(m => identifierUtils.makePathsRelative(compiler.context, m.identifier(), compilation.cache));
 					if(typeof chunk._fromAggressiveSplittingIndex === "undefined") {
 						if(incorrectSize) return;
 						chunk.recorded = true;