--- a/lib/util/identifier.js
+++ b/lib/util/identifier.js
@@ -7,10 +7,18 @@ const looksLikeAbsolutePath = (maybeAbsolutePath) => {
 
 const normalizePathSeparator = (p) => p.replace(/\\/g, "/");
 
-/* A map from context dir strings to maps from identifier names to relative paths */
-const relativePaths = new Map();
+const _makePathsRelative = (context, identifier) => {
+	return identifier
+		.split(/([|! ])/)
+		.map(str => looksLikeAbsolutePath(str) ?
+			normalizePathSeparator(path.relative(context, str)) : str)
+		.join("");
+}
 
-exports.makePathsRelative = (context, identifier) => {
+exports.makePathsRelative = (context, identifier, cache) => {
+	if(!cache) return _makePathsRelative(context, identifier);
+
+	const relativePaths = cache.relativePaths || (cache.relativePaths = new Map());
 	if(!relativePaths.has(context)) {
 		relativePaths.set(context, new Map());
 	}
@@ -20,11 +28,7 @@ exports.makePathsRelative = (context, identifier) => {
 	if(contextCache.has(identifier)) {
 		return contextCache.get(identifier);
 	} else {
-		var relativePath = identifier
-			.split(/([|! ])/)
-			.map(str => looksLikeAbsolutePath(str) ?
-				normalizePathSeparator(path.relative(context, str)) : str)
-			.join("");
+		let relativePath = _makePathsRelative(context, identifier);
 		contextCache.set(identifier, relativePath);
 		return relativePath;
 	}