--- a/lib/util/identifier.js
+++ b/lib/util/identifier.js
@@ -13,7 +13,7 @@ const _makePathsRelative = (context, identifier) => {
 		.map(str => looksLikeAbsolutePath(str) ?
 			normalizePathSeparator(path.relative(context, str)) : str)
 		.join("");
-}
+};
 
 exports.makePathsRelative = (context, identifier, cache) => {
 	if(!cache) return _makePathsRelative(context, identifier);