--- a/lib/util/identifier.js
+++ b/lib/util/identifier.js
@@ -19,18 +19,20 @@ exports.makePathsRelative = (context, identifier, cache) => {
 	if(!cache) return _makePathsRelative(context, identifier);
 
 	const relativePaths = cache.relativePaths || (cache.relativePaths = new Map());
-	if(!relativePaths.has(context)) {
-		relativePaths.set(context, new Map());
+
+	let cachedResult;
+	let contextCache = relativePaths.get(context)
+	if(typeof contextCache === "undefined") {
+		relativePaths.set(context, contextCache = new Map());
+	} else {
+		cachedResult = contextCache.get(identifier)
 	}
 
-	const contextCache = relativePaths.get(context);
-
-	if(contextCache.has(identifier)) {
-		return contextCache.get(identifier);
+	if(typeof cachedResult !== "undefined") {
+		return cachedResult;
 	} else {
-		let relativePath = _makePathsRelative(context, identifier);
+		const relativePath = _makePathsRelative(context, identifier);
 		contextCache.set(identifier, relativePath);
 		return relativePath;
 	}
-
 };