--- a/lib/util/identifier.js
+++ b/lib/util/identifier.js
@@ -21,11 +21,11 @@ exports.makePathsRelative = (context, identifier, cache) => {
 	const relativePaths = cache.relativePaths || (cache.relativePaths = new Map());
 
 	let cachedResult;
-	let contextCache = relativePaths.get(context)
+	let contextCache = relativePaths.get(context);
 	if(typeof contextCache === "undefined") {
 		relativePaths.set(context, contextCache = new Map());
 	} else {
-		cachedResult = contextCache.get(identifier)
+		cachedResult = contextCache.get(identifier);
 	}
 
 	if(typeof cachedResult !== "undefined") {