@asolove 2017-07-17
I've addressed the code review comments by splitting the non-memoized implementation into its own function and storing the cache on the `compilation` object so it can be GC'd.