@asolove 2017-07-18
I _think_ the failing code coverage comparison check was because my branch wasn't up-to-date with master, so I've rebased to try to get it green.