@sokra 2017-07-18
I made a small change to the code, trying to reduce the number of calls to `Map`. I called `get` directly without checking for `has`.